+++ lib/auth/github_test.go
@@ -43,7 +43,6 @@ import (
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/backend"
 	"github.com/gravitational/teleport/lib/backend/memory"
-	"github.com/gravitational/teleport/lib/client/sso"
 	"github.com/gravitational/teleport/lib/events"
 	"github.com/gravitational/teleport/lib/events/eventstest"
 	"github.com/gravitational/teleport/lib/loginrule"
@@ -610,198 +609,4 @@ func TestBuildAPIEndpoint(t *testing.T) {
 			require.Equal(t, tt.expected, got)
 		})
 	}
-}
-
-func TestValidateClientRedirect(t *testing.T) {
-	t.Run("StandardLocalhost", func(t *testing.T) {
-		for _, goodURL := range []string{
-			"http://127.0.0.1/callback",
-			"http://127.0.0.1:1234/callback",
-			"http://[::1]/callback",
-			"http://[::1]:1234/callback",
-			"http://localhost/callback",
-			"http://localhost:1234/callback",
-		} {
-			var defaultSettings *types.SSOClientRedirectSettings
-			require.NoError(t, sso.ValidateClientRedirect(goodURL+"?secret_key=", sso.CeremonyTypeLogin, defaultSettings))
-		}
-	})
-
-	t.Run("InvalidLocalhostLike", func(t *testing.T) {
-		for _, badURL := range []string{
-			"http://127.0.0.1:12345/notcallback",
-			"http://127a0.0.1/callback",
-			"http://127a0.0.1:1234/callback",
-			"http://::1/callback",
-			"http://notlocalhost/callback",
-			"http://sub.localhost/callback",
-			"http://localhost.com/callback",
-			"http://127.0.0.1.example.com/callback",
-			"http://[::1].example.com/callback",
-			"http://username@127.0.0.1:12345/callback",
-			"http://@localhost:12345/callback",
-			"http://localhost@example.com/callback",
-			"http://127.0.0.1:12345@example.com/callback",
-			"https://127.0.0.1:12345/callback",
-			"https://localhost:12345/callback",
-			"https://localhost/callback",
-			"ftp://localhost/callback",
-			"ftp://127.0.0.1/callback",
-			"ftp://[::1]/callback",
-		} {
-			var defaultSettings *types.SSOClientRedirectSettings
-			require.Error(t, sso.ValidateClientRedirect(badURL+"?secret_key=", sso.CeremonyTypeLogin, defaultSettings))
-		}
-	})
-
-	t.Run("BadQuery", func(t *testing.T) {
-		for _, badURL := range []string{
-			"http://127.0.0.1:12345/callback",
-			"http://127.0.0.1:12345/callback?secret=a",
-			"http://127.0.0.1:12345/callback?secret_key=a&foo=b",
-		} {
-			var defaultSettings *types.SSOClientRedirectSettings
-			require.Error(t, sso.ValidateClientRedirect(badURL, sso.CeremonyTypeLogin, defaultSettings))
-		}
-	})
-
-	t.Run("AllowedHttpsHostnames", func(t *testing.T) {
-		for _, goodURL := range []string{
-			"https://allowed.domain.invalid/callback",
-			"https://foo.allowed.with.subdomain.invalid/callback",
-			"https://but.no.subsubdomain.invalid/callback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				AllowedHttpsHostnames: []string{
-					"allowed.domain.invalid",
-					"*.allowed.with.subdomain.invalid",
-					"^[-a-zA-Z0-9]+.no.subsubdomain.invalid$",
-				},
-			}
-			require.NoError(t, sso.ValidateClientRedirect(goodURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-
-		for _, badURL := range []string{
-			"https://allowed.domain.invalid/notcallback",
-			"https://allowed.domain.invalid:12345/callback",
-			"http://allowed.domain.invalid/callback",
-			"https://not.allowed.domain.invalid/callback",
-			"https://allowed.with.subdomain.invalid/callback",
-			"https://i.said.no.subsubdomain.invalid/callback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				AllowedHttpsHostnames: []string{
-					"allowed.domain.invalid",
-					"*.allowed.with.subdomain.invalid",
-					"^[-a-zA-Z0-9]+.no.subsubdomain.invalid",
-				},
-			}
-			require.Error(t, sso.ValidateClientRedirect(badURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-	})
-
-	t.Run("InsecureAllowedCidrRanges", func(t *testing.T) {
-		for _, goodURL := range []string{
-			"http://192.168.0.27/callback",
-			"https://192.168.0.27/callback",
-			"http://192.168.0.27:1337/callback",
-			"https://192.168.0.27:1337/callback",
-			"http://[2001:db8::aaaa:bbbb]/callback",
-			"https://[2001:db8::aaaa:bbbb]/callback",
-			"http://[2001:db8::aaaa:bbbb]:1337/callback",
-			"https://[2001:db8::aaaa:bbbb]:1337/callback",
-			"http://[2001:db8::1]/callback",
-			"https://[2001:db8::1]/callback",
-			"http://[2001:db8::1]:1337/callback",
-			"https://[2001:db8::1]:1337/callback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				InsecureAllowedCidrRanges: []string{
-					"192.168.0.0/24",
-					"2001:db8::/96",
-				},
-			}
-			require.NoError(t, sso.ValidateClientRedirect(goodURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-
-		for _, badURL := range []string{
-			"http://192.168.1.1/callback",
-			"https://192.168.1.1/callback",
-			"http://192.168.1.1:80/callback",
-			"https://192.168.1.1:443/callback",
-			"http://[2001:db8::1:aaaa:bbbb]/callback",
-			"https://[2001:db8::1:aaaa:bbbb]/callback",
-			"http://[2001:db8::1:aaaa:bbbb]:80/callback",
-			"https://[2001:db8::1:aaaa:bbbb]:443/callback",
-			"http://[2001:db9::]/callback",
-			"https://[2001:db9::]/callback",
-			"http://not.an.ip/callback",
-			"https://not.an.ip/callback",
-			"http://192.168.0.27/nocallback",
-			"https://192.168.0.27/nocallback",
-			"http://[2001:db8::1]/notacallback",
-			"https://[2001:db8::1]/notacallback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				InsecureAllowedCidrRanges: []string{
-					"192.168.0.0/24",
-					"2001:db8::/96",
-				},
-			}
-			require.Error(t, sso.ValidateClientRedirect(badURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-	})
-
-	t.Run("SSOTestFlow", func(t *testing.T) {
-		for _, goodURL := range []string{
-			"http://127.0.0.1:12345/callback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				AllowedHttpsHostnames: []string{
-					"allowed.domain.invalid",
-				},
-			}
-			require.NoError(t, sso.ValidateClientRedirect(goodURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-
-		for _, badURL := range []string{
-			"https://allowed.domain.invalid/callback",
-			"http://allowed.domain.invalid/callback",
-		} {
-			settings := &types.SSOClientRedirectSettings{
-				AllowedHttpsHostnames: []string{
-					"allowed.domain.invalid",
-				},
-			}
-			require.Error(t, sso.ValidateClientRedirect(badURL+"?secret_key=", sso.CeremonyTypeLogin, settings))
-		}
-	})
-
-	t.Run("SSOMFAWeb", func(t *testing.T) {
-		settings := &types.SSOClientRedirectSettings{
-			AllowedHttpsHostnames: []string{
-				"allowed.domain.invalid",
-			},
-		}
-
-		// Only allow web mfa redirect as a relative path.
-		require.NoError(t, sso.ValidateClientRedirect(sso.WebMFARedirect+"?channel_id=", sso.CeremonyTypeLogin, settings))
-
-		for _, badURL := range []string{
-			"localhost:12345" + sso.WebMFARedirect,
-			"http://localhost:12345" + sso.WebMFARedirect,
-			"https://localhost:12345" + sso.WebMFARedirect,
-			"127.0.0.1:12345" + sso.WebMFARedirect,
-			"http://127.0.0.1:12345" + sso.WebMFARedirect,
-			"https://127.0.0.1:12345" + sso.WebMFARedirect,
-			"allowed.domain.invalid" + sso.WebMFARedirect,
-			"http://allowed.domain.invalid" + sso.WebMFARedirect,
-			"https://allowed.domain.invalid" + sso.WebMFARedirect,
-			"not.allowed.domain.invalid" + sso.WebMFARedirect,
-			"http://not.allowed.domain.invalid" + sso.WebMFARedirect,
-			"https://not.allowed.domain.invalid" + sso.WebMFARedirect,
-		} {
-			require.Error(t, sso.ValidateClientRedirect(badURL+"?channel_id=", sso.CeremonyTypeLogin, settings))
-		}
-	})
 }
